refactor(desktop): move GIF stripping to commands/media_gif.rs
The GIF sanitizer pushed media.rs past the 1000-line desktop file-size
limit (check-file-sizes.mjs). Split it into its own module, following the
existing media_transcode/media_download precedent. No behavior change.
