+++ desktop/src-tauri/src/commands/mod.rs
@@ -24,6 +24,7 @@ mod legacy_storage;
 mod link_preview;
 pub(crate) mod media;
 mod media_download;
+mod media_gif;
 mod media_transcode;
 #[cfg(feature = "mesh-llm")]
 pub(crate) mod mesh_llm;
